feat: 각 폴더 분리, (커스텀 memo함수에 오류 발견...)
